Added test to credentials

+++ app/src/main/java/com/example/ibm_test/data/UserInfoData.kt
@@ -12,5 +12,5 @@ data class UserInfoData (
     @SerializedName("name") val name: String,
     @SerializedName("bankAccount") val bankAccount: String,
     @SerializedName("agency") val agency: String,
-    @SerializedName("balance") val balance: Float
+    @SerializedName("balance") val balance: Double
 ) : Serializable
